Use TRUFFLERUBY_CEXT_ENABLED to decide whether to build C extensions
* We no longer need SULONG_HOME to build C extensions.
* This mostly reverts commit d42bda31dd3e3df8337e22ffc276662030e79016.

+++ lib/patches/stdlib/rubygems/ext.rb
@@ -1,35 +1,35 @@
 Truffle::Patching.require_original __FILE__
 
-# TruffleRuby: Ignore native extensions
+# TruffleRuby: build C extensions conditionally
 
 class Gem::Ext::Builder
   def build_extensions
     return if @spec.extensions.empty?
 
-    if @build_args.empty?
-      say "Building native extensions.  This could take a while..."
-    else
-      say "Building native extensions with: '#{@build_args.join ' '}'"
-      say "This could take a while..."
-    end
-
-    dest_path = @spec.extension_dir
-
-    FileUtils.rm_f @spec.gem_build_complete_path
-
-    @ran_rake = false # only run rake once
-
-    sulong_present = !!ENV['SULONG_HOME']
-    @spec.extensions.each do |extension|
-      break if @ran_rake
-      # TruffleRuby: build C extensions conditionally
-      if sulong_present
-        build_extension extension, dest_path
+    if ENV['TRUFFLERUBY_CEXT_ENABLED']
+      if @build_args.empty?
+        say "Building native extensions.  This could take a while..."
       else
-        puts 'SULONG_HOME is not set therefore skipping C extension build.'
+        say "Building native extensions with: '#{@build_args.join ' '}'"
+        say "This could take a while..."
       end
-    end
 
-    FileUtils.touch @spec.gem_build_complete_path if sulong_present
+      dest_path = @spec.extension_dir
+
+      FileUtils.rm_f @spec.gem_build_complete_path
+
+      @ran_rake = false # only run rake once
+
+      @spec.extensions.each do |extension|
+        break if @ran_rake
+
+        build_extension extension, dest_path
+      end
+
+      FileUtils.touch @spec.gem_build_complete_path
+    else
+      puts "WORKAROUND: Not building native extensions for #{@spec.name}.\n" +
+           'Support of C extensions is in development, set TRUFFLERUBY_CEXT_ENABLED=true to experiment.'
+    end
   end
 end
